add product edit and show views

+++ db/migrate/20160219051242_create_measurements.rb
@@ -4,11 +4,11 @@ class CreateMeasurements < ActiveRecord::Migration
         t.integer :measureable_id
         t.string :measureable_type
     	t.string :gender
-    	t.integer :hip
-    	t.integer :waist
-    	t.integer :bust
-    	t.integer :chest
-    	t.integer :inseam
+    	t.float :hip
+    	t.float :waist
+    	t.float :bust
+    	t.float :chest
+    	t.float :inseam
     	t.timestamps null: false
     end
   end
